backend/kafka/producer/requirements.txt
kafka-python==2.0.2
neo4j==5.27.0
spotipy==2.24.0
tqdm==4.67.1
lrclibapi==0.3.1
llama-cloud==0.1.6
llama-index==0.12.8
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.8
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.1
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.18
psycopg2-binary==2.9.10

